Move new merge_experiment method to the group file.
